+++ code/alphashp.cpp
@@ -259,12 +259,12 @@ void AlphaShapeClass::Draw_In_Area(Point2D const & point, Rect const & cliprect)
 								*alphaptr = BrightnessTable[pixel][*alphaptr];
 							}
 							alphaptr++;
-							alphaptr = (unsigned short *)AlphaBuffer->Wrap_Overflow((unsigned)alphaptr);
+							alphaptr = (unsigned short *)AlphaBuffer->Wrap_Overflow((uintptr_t)alphaptr);
 						}
 						shapeptr += shape_skip;
 						maskptr += mask_skip;
 						alphaptr += alpha_skip;
-						alphaptr = (unsigned short *)AlphaBuffer->Wrap_Overflow((unsigned)alphaptr);
+						alphaptr = (unsigned short *)AlphaBuffer->Wrap_Overflow((uintptr_t)alphaptr);
 					}
 				} else {
 					for (int i = top; i < bottom; i++) {
@@ -336,11 +336,11 @@ void AlphaShapeClass::Draw_All(Rect const & cliprect)
 							unsigned char pixel = *shapeptr++;
 							*alphaptr = BrightnessTable[pixel][*alphaptr];
 							alphaptr++;
-							alphaptr = (unsigned short *)AlphaBuffer->Wrap_Overflow((unsigned)alphaptr);
+							alphaptr = (unsigned short *)AlphaBuffer->Wrap_Overflow((uintptr_t)alphaptr);
 						}
 						shapeptr += shape_skip;
 						alphaptr += alpha_skip;
-						alphaptr = (unsigned short *)AlphaBuffer->Wrap_Overflow((unsigned)alphaptr);
+						alphaptr = (unsigned short *)AlphaBuffer->Wrap_Overflow((uintptr_t)alphaptr);
 					}
 				} else {
 					for (int i = top; i < bottom; i++) {
